fix too many errors

+++ code/src/bilist.cc
@@ -2,13 +2,13 @@
 
 namespace bilsmtree {
 
-BiList::BiList(uint size) {
+BiList::BiList(size_t size) {
   head_ = 0;
   tail_ = 0;
   data_size_ = 0;
   max_size_ = size;
   data_ = new ListNode[size + 1];
-  for (uint i = 1; i <= size; ++ i)
+  for (size_t i = 1; i <= size; ++ i)
     free_.push(i);
 }
 
@@ -18,12 +18,12 @@ BiList::~BiList() {
   delete data_;
 }
 
-void BiList::Set(uint pos, const KV kv) {
+void BiList::Set(size_t pos, const KV kv) {
   Util::Assert("Position is exceeded!\nMethod: BiList::Set", (pos > 0 && pos <= max_size_));
   data_[pos].kv_ = kv;
 }
 
-void BiList::MoveToHead(uint pos) {
+void BiList::MoveToHead(size_t pos) {
   Util::Assert("Position is exceeded!\nMethod: BiList::MoveToHead", (pos > 0 && pos <= max_size_));
   if (pos == tail_) {
     tail_ = data_[tail_].prev_;
@@ -35,12 +35,12 @@ void BiList::MoveToHead(uint pos) {
   data_[data_[pos].next_].prev_ = pos;
 }
 
-KV BiList::Get(uint pos) {
+KV BiList::Get(size_t pos) {
   Util::Assert("Position is exceeded!\nMethod: BiList::Get", (pos > 0 && pos <= max_size_));
   return data_[pos].kv_;
 }
 
-KV BiList::Delete(uint pos) {
+KV BiList::Delete(size_t pos) {
   Util::Assert("Position is exceeded!\nMethod: BiList::Delete", (pos > 0 && pos <= max_size_));
   KV kv = data_[pos].kv_;
   data_[data_[pos].prev_].next_ = data_[pos].next_;
@@ -59,7 +59,7 @@ bool BiList::Append(const KV kv, KV& pop_kv) {
     pop_kv = Delete(tail_);
     res = true;
   }
-  uint pos = free_.front();
+  size_t pos = free_.front();
   free_.pop();
   data_size_ = data_size_ + 1;
   data_[pos].kv_ = kv;
@@ -76,7 +76,7 @@ bool BiList::Insert(const KV kv, KV& pop_kv) {
     pop_kv = Delete(data_[head_].next_);
     res = true;
   }
-  uint pos = free_.front();
+  size_t pos = free_.front();
   free_.pop();
   data_size_ = data_size_ + 1;
   data_[pos].kv_ = kv;
