+++ code/src/hash.cc
@@ -2,16 +2,16 @@
 
 namespace bilsmtree {
 
-uint Hash(const char* data, uint n, size_t seed) {
+size_t Hash(const char* data, size_t n, size_t seed) {
   // Similar to murmur hash
   const size_t m = 0xc6a4a793;
-  const uint r = 24;
+  const size_t r = 24;
   const char* limit = data + n;
   size_t h = seed ^ (n * m);
 
   // Pick up four bytes at a time
   while (data + 4 <= limit) {
-    uint w;
+    size_t w;
     memcpy(&w, data, sizeof(w));  // gcc optimizes this to a plain load
     data += 4;
     h += w;
